fixed problem with first load of cache, and bug where row is created in bottom of page.

+++ core/src/main/java/com/thundashop/core/socket/CacheFactory.java
@@ -187,8 +187,10 @@ public class CacheFactory {
             }
         });
 
-        for (final File file : files) {
-            deleteDirectory(file);
+        if(files != null ) {
+            for (final File file : files) {
+                deleteDirectory(file);
+            }
         }
     }
